# 05-agents/openapi.yaml
openapi: 3.0.3
info:
  title: LangChain4j Agent Tools API
  description: OpenAPI specification for tools exposed to Azure AI Agent Service
  version: 1.0.0
  contact:
    name: Agent Tools API
    url: https://github.com/roryp/langchainagent

servers:
  - url: https://ca-agents-{resourceToken}.{region}.azurecontainerapps.io
    description: Azure Container App deployment
    variables:
      resourceToken:
        default: ozhotol5yje76
      region:
        default: livelyplant-8c187f15.eastus2
  - url: http://localhost:8082
    description: Local development server

paths:
  /api/tools/weather/current:
    post:
      operationId: getCurrentWeather
      summary: Get current weather for a location
      description: Retrieves the current weather conditions for a specified location
      tags:
        - Weather
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - location
              properties:
                location:
                  type: string
                  description: The location to get weather for (e.g., "Seattle", "London, UK")
                  example: "Seattle"
      responses:
        '200':
          description: Weather information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                  temperature:
                    type: integer
                    description: Temperature in Celsius
                  condition:
                    type: string
                    description: Weather condition
                  description:
                    type: string
                    description: Full weather description
        '400':
          description: Invalid request
        '500':
          description: Internal server error

  /api/tools/weather/forecast:
    post:
      operationId: getWeatherForecast
      summary: Get weather forecast
      description: Retrieves weather forecast for a specified location and number of days
      tags:
        - Weather
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - location
                - days
              properties:
                location:
                  type: string
                  description: The location to get forecast for
                  example: "Seattle"
                days:
                  type: integer
                  description: Number of days (1-7)
                  minimum: 1
                  maximum: 7
                  example: 3
      responses:
        '200':
          description: Forecast retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                  days:
                    type: integer
                  forecast:
                    type: array
                    items:
                      type: object
                      properties:
                        day:
                          type: integer
                        temperature:
                          type: integer
                        condition:
                          type: string
        '400':
          description: Invalid request
        '500':
          description: Internal server error

  /api/tools/calculator/add:
    post:
      operationId: add
      summary: Add two numbers
      description: Calculates the sum of two numbers
      tags:
        - Calculator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - a
                - b
              properties:
                a:
                  type: number
                  description: First number
                  example: 25
                b:
                  type: number
                  description: Second number
                  example: 17
      responses:
        '200':
          description: Calculation successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: number
                  operation:
                    type: string

  /api/tools/calculator/subtract:
    post:
      operationId: subtract
      summary: Subtract two numbers
      description: Calculates the difference between two numbers
      tags:
        - Calculator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - a
                - b
              properties:
                a:
                  type: number
                  description: Minuend (number to subtract from)
                b:
                  type: number
                  description: Subtrahend (number to subtract)
      responses:
        '200':
          description: Calculation successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: number

  /api/tools/calculator/multiply:
    post:
      operationId: multiply
      summary: Multiply two numbers
      description: Calculates the product of two numbers
      tags:
        - Calculator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - a
                - b
              properties:
                a:
                  type: number
                b:
                  type: number
      responses:
        '200':
          description: Calculation successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: number

  /api/tools/calculator/divide:
    post:
      operationId: divide
      summary: Divide two numbers
      description: Calculates the quotient of two numbers
      tags:
        - Calculator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - a
                - b
              properties:
                a:
                  type: number
                  description: Dividend
                b:
                  type: number
                  description: Divisor (cannot be zero)
      responses:
        '200':
          description: Calculation successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: number
        '400':
          description: Division by zero

  /api/tools/calculator/power:
    post:
      operationId: power
      summary: Calculate power
      description: Calculates base raised to exponent
      tags:
        - Calculator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - base
                - exponent
              properties:
                base:
                  type: number
                  description: Base number
                exponent:
                  type: number
                  description: Exponent
      responses:
        '200':
          description: Calculation successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: number

  /api/tools/calculator/sqrt:
    post:
      operationId: squareRoot
      summary: Calculate square root
      description: Calculates the square root of a number
      tags:
        - Calculator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - number
              properties:
                number:
                  type: number
                  description: Number to calculate square root of (must be non-negative)
                  minimum: 0
      responses:
        '200':
          description: Calculation successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: number
        '400':
          description: Negative number provided

  /health:
    get:
      operationId: healthCheck
      summary: Health check endpoint
      description: Returns the health status of the tools service
      tags:
        - System
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  service:
                    type: string
                  timestamp:
                    type: integer

components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: integer